+++ client/Packages/com.beamable/CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Content deletion popup opens as separate window
 - An optional `forceRefresh` parameter to all subscribable SDK's `GetCurrent()` method that forces a network request
+- `API.Instance.Requester` is now an `IBeamableRequester`
 
 ### Fixed
 - If no internet connection exists on startup, `API.Instance()` will retry every 2 seconds until a connection is established